\name{buildgraph}
\alias{buildgraph}
\title{buildgraphs}
\usage{
useful to build graphs
}
\description{
The package contains a function that takes a dataset, the variable indexes and the directory location as arguments and creates bar plots and pie charts for categorical  variables and box plots and histograms for numerical variables.
}
\examples{
buildgraphs(data,var,dir)
}
